Move the login and api service into api folder
