+++ src/app/services/PeerlyServices/api/index.ts
@@ -1,6 +1,7 @@
 import axios, {AxiosInstance, AxiosRequestConfig} from 'axios';
 import Config from 'react-native-config';
-import AsyncStore from '../asyncStorage';
+import AsyncStore from '../../asyncStorage';
+import PeerlyAsyncStore from '../peerlyAsyncStorage';
 
 export const axiosInstance: AxiosInstance = axios.create({
   baseURL: Config.PEERLY_API_BASE_URL,
@@ -10,8 +11,7 @@ export const axiosInstance: AxiosInstance = axios.create({
   },
 });
 
-import {PEERLY_LOGIN_ROUTE} from '../../constant/apiRoutes';
-import {PeerlyLoginResponse} from './types';
+import {PEERLY_LOGIN_ROUTE} from '../../../screens/PeerlyScreens/constants/apiRoutes';
 
 export const apiCall = async <T, D>(config: AxiosRequestConfig<T>) => {
   let authToken: string | null;
@@ -21,7 +21,9 @@ export const apiCall = async <T, D>(config: AxiosRequestConfig<T>) => {
       'Intranet-Auth': authToken,
     };
   } else {
-    authToken = await AsyncStore.getItem(AsyncStore.PEERLY_AUTH_TOKEN_KEY);
+    authToken = await PeerlyAsyncStore.getItem(
+      PeerlyAsyncStore.PEERLY_AUTH_TOKEN_KEY,
+    );
     const authorizationHeader = authToken;
 
     if (config.headers) {
@@ -34,12 +36,4 @@ export const apiCall = async <T, D>(config: AxiosRequestConfig<T>) => {
   }
   const response = await axiosInstance.request<D>(config);
   return response;
-};
-
-export const loginPeerly = async () => {
-  const response = await apiCall<any, PeerlyLoginResponse>({
-    method: 'GET',
-    url: PEERLY_LOGIN_ROUTE,
-  });
-  return response.data;
 };
